+++ Victorious/Tournament.Structure/Classes/SingleElimBracket.cs
@@ -140,7 +140,7 @@ namespace Tournament.Structure
 		#endregion
 
 		#region Public Methods
-		public override void CreateBracket(ushort _gamesPerMatch = 1)
+		public override void CreateBracket(int _gamesPerMatch = 1)
 		{
 			if (_gamesPerMatch < 1)
 			{
